refactor: extract databend-meta-snapshot-store crate
# Summary

Snapshot files — building them, streaming them between nodes, and loading
them back — become their own crate, a sibling of the state machine rather
than a layer under it.

# Details

The two exchange `DB` handles from `snapshot-db` and neither imports the
other; `service` is what wires them together. That independence is now
enforced by cargo rather than by convention.

`raft-store` keeps only the on-disk upgrader's use of the snapshot store,
which is why it still depends on the new crate while no longer depending
on the state machine.
