+++ crates/server/service/Cargo.toml
@@ -25,6 +25,7 @@ databend-meta-raft-store = { workspace = true }
 databend-meta-runtime-api = { workspace = true }
 databend-meta-sled-store = { workspace = true }
 databend-meta-snapshot-db = { workspace = true }
+databend-meta-snapshot-store = { workspace = true }
 databend-meta-state-machine = { workspace = true }
 databend-meta-types = { workspace = true }
 databend-meta-version = { workspace = true }
