+++ crates/server/service/src/meta_service/raft_service_impl.rs
@@ -23,14 +23,14 @@ use databend_meta_client::MetaGrpcReadReq;
 use databend_meta_leveled_store::persisted_codec::PersistedCodec;
 use databend_meta_raft_config::data_version::DATA_VERSION;
 use databend_meta_raft_config::snapshot_config::SnapshotConfig;
-use databend_meta_raft_store::snapshot_store::MetaSnapshotId;
-use databend_meta_raft_store::snapshot_store::SnapshotWriter;
-use databend_meta_raft_store::snapshot_store::open_snapshot::OpenSnapshot;
-use databend_meta_raft_store::snapshot_store::received::Received;
-use databend_meta_raft_store::snapshot_store::write_entry::WriteEntry;
 use databend_meta_runtime_api::SpawnApi;
 use databend_meta_snapshot_db::DB;
 use databend_meta_snapshot_db::Snapshot;
+use databend_meta_snapshot_store::MetaSnapshotId;
+use databend_meta_snapshot_store::SnapshotWriter;
+use databend_meta_snapshot_store::open_snapshot::OpenSnapshot;
+use databend_meta_snapshot_store::received::Received;
+use databend_meta_snapshot_store::write_entry::WriteEntry;
 use databend_meta_types::GrpcHelper;
 use databend_meta_types::MetaAPIError;
 use databend_meta_types::PbAppendRequestExt;
