+++ crates/server/service/src/store/store.rs
@@ -30,11 +30,11 @@ use databend_meta_raft_store::db_exporter::DBExporter;
 use databend_meta_raft_store::header::Header;
 use databend_meta_raft_store::ondisk::TREE_HEADER;
 use databend_meta_raft_store::sled_compat::key_spaces::RaftStoreEntry;
-use databend_meta_raft_store::snapshot_store::MetaSnapshotId;
-use databend_meta_raft_store::snapshot_store::SnapshotStore;
 use databend_meta_runtime_api::SpawnApi;
 use databend_meta_snapshot_db::DB;
 use databend_meta_snapshot_db::DBStat;
+use databend_meta_snapshot_store::MetaSnapshotId;
+use databend_meta_snapshot_store::SnapshotStore;
 use databend_meta_state_machine::StateMachine;
 use databend_meta_types::Endpoint;
 use databend_meta_types::Node;
