+++ crates/server/snapshot-store/src/db_open_snapshot_impl.rs
@@ -22,7 +22,7 @@ use openraft::SnapshotId;
 use rotbl::storage::impls::fs::FsStorage;
 use rotbl::v001::Rotbl;
 
-use crate::snapshot_store::open_snapshot::OpenSnapshot;
+use crate::open_snapshot::OpenSnapshot;
 
 impl OpenSnapshot for DB {
     fn open_snapshot(
