+++ crates/server/snapshot-store/src/open_snapshot.rs
