+++ crates/server/snapshot-store/src/snapshot_store_error.rs
