+++ crates/server/snapshot-store/src/write_entry.rs
