+++ crates/server/snapshot-store/src/writer_stat.rs
